+++ contrib/natss/pkg/dispatcher/channel/controller.go
@@ -24,7 +24,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/source"
 
 	eventingv1alpha1 "github.com/knative/eventing/pkg/apis/eventing/v1alpha1"
-	"github.com/knative/eventing/pkg/provisioners/natss/dispatcher/dispatcher"
+	"github.com/knative/eventing/contrib/natss/pkg/dispatcher/dispatcher"
 )
 
 const (
